Embedded mysql fixes.
There is now a mysql_server_init() function which needs to be
called at the beginning of the program (and _end() for the end
of the program).  This routine handles argument parsing for the
embedded server.

Use the embedded version of mysql_load_file() (ignore the LOCAL
argument, since the client and server are the same program).

There are now mysql_thread_init/end() functions for the client
to use in a multi-threaded app.  They are just wrappers for
my_thread_init/end().