Task 430: Allowing braces in joins by simply removing them.
This is a simple fix, allowing a join_table_list in the right reduction of a
normal_join sequence, instead of just a join_table. This makes things like
"t1, (t2 left join t3)" work, but it also allows "join" and "cross join" instead
of ",".

This should fix the bug reported as:
  Subject: ODBC SQL syntax issue 
  From: Ivan Vazharov 
  Date: Mon, 30 Sep 2002 12:02:42 +0200