Bug #5185  	mysqldump for windows database gets table names with different case
Added the get_actual_table_name function that issues a SHOW TABLES LIKE '%s'.  This will get the table 
name in the proper case.  We use this table name rather than the one given on the command line.  This will
prevent problems when importing SQL on Linux that was generated on a Windows platform where case can be
an issue.

mysqldump.c:
  call get_actual_table_name to get the table name in the proper case