WL #2094 Federated Storage Engine
Changes to a test with a update to a table with 1001 columns, the last being a blob. 
On HP, the test worked properly, but on other OSs, there's a problem with using the 
blob column in an update where clause. The field method val_string(String *, char *) 
should work, but doesn't seem to deal with blobd, and for some reason, cannot get
the correct value of the blob column from the byte pointer in update_row, "*old_data"
which is a byte pointer to the row in mysql format, that will be replaced with 
"*new_data" (which conveniently has a field pointer to use to get values from).

Will document this.