Bug#18470: Compile Failure: strings.s fails make with unknown pseudo-ops
The bug is thqt we use some nonstandard assembly codes in our strings
source, and some assemblers don't know what to do with them:  Specifically,
Sun's Solaris assembler and Apple's Darwin assembler balk at them.

This patch, rather than trying to test for properties of the assembler, 
which Autoconf doesn't have any decent facilites for, instead tries to
compile the code in question and disables assembly if it fails.

There's still the problem of unportable assembly, but I'll leave that 
to someone who feels like rewriting and debugging it.