Bug #19148  	Backup causes cluster down if _really_ early abort happens
Early abort, failing to allocate buffers, would go down the abort track and
 end up trying to sendDropTriggers (as the standard stop backup does). 
However, it is possible to not have any tables yet defined in the backup 
(due to allocating buffers is before this). Simple check in sendDropTrig 
(and continue to next step). Files may be open, so we continue to close them.

Also updated ERROR_Codes (and added one to test this bug).